Create NubsScript v2.2.js
This is my script for Nanogames.io that I have build from scratch. I do not claim to be an expert, just an avid learner. There are many ideas I would like to add to this script. I have a long journey ahead of me. GitHub realizes collaboration spawns innovation and it is time for me to see what it can do for me